Move frontend pages to nested directory
